// apps/web/src/lib/theme-palettes.generated.ts
export interface ThemePalette {
  id: string;
  name: string;
  type: "light" | "dark";
  bg: string;
  fg: string;
  keyword: string;
  string: string;
  func: string;
  comment: string;
  variable: string;
  number: string;
}